Implemented the ability to pre-render subs.
Upside: draw() is simply a cloneNode() of the pre-rendered div, so it's faster.
Downside: Each Dialogue now holds a reference to a DOM element, increasing memory usage.

Perhaps a future improvement can use a heuristic of how complex the Dialogue is to render, so that only those Dialogue's that take time to render are pre-rendered and saved, while simpler Dialogue's are re-rendered everytime they're displayed.