Fixed font sizes.
The font size in ASS is actually meant to be the line height. ASS renderers scale the font size so that it fits in the line height. Since there is no way in the browser to calculate a font size such that it results in a given line height, libjass now uses an invisible div's offsetHeight to calculate the line height for a given font size, and works backwards from there to approximate the font size required for a target line height.