Added Ping WorkerCommand to fix random failure in web worker tests.
The Worker constructor returns before the web worker is ready to respond to messages, and even before its script has been downloaded. This was causing problems with the web worker test because it would take more than the individual test's timeout for the worker to become ready.

This went unnoticed before because the code to create the worker ran as soon as the suite was loaded by the test runner (as opposed to when it was executed), so the worker would begin being created then. Since the web worker suite ran last, by the time the other suites finished running the worker would actually be ready and the tests would have no problem.

However this stopped being true when the code to create the worker channel for the suite was moved to its before() function, since now it would run when the suite was being executed. As a result, the first test would time out while waiting for the worker to be created.