Calculate font metrics while preloading fonts.
The sync method of setting the fontSize div to the requested font family and measuring its offsetHeight doesn't work when the request font family is from a @font-face declaration. Atleast some browsers render the div with text in a locally installed fallback font and update it to the actual font in a future tick.

For example, if the metrics are being calculated for `Foo, Arial, Helvetica, sans-serif, 'Segoe UI Symbol'` where Foo is defined via a `@font-face` declaration, then the M glyph is rendered in Arial for that tick, and updated to Foo in a future tick.

This change adds a delay between setting the font on the div and measuring its offsetHeight, which requires making the metric calculation async. Since the renderer pipeline cannot be made async, this change also causes the font metric to be computed while the font is being preloaded, since preloading is already an async operation. The original sync method still remains to be able to calculate metrics for non-preloaded fonts; a note has been added that this will result in wrong sizes for these fonts unless they're already installed on the user's machine.

Reported in #55