Added support for using attached fonts.
Attached fonts are loaded as data: URIs. Since ASS does not contain font family names for attached fonts (only the original filename which is useless), libjass will parse the font as a TTF to extract the font names from the name tables. The TTF parser is simple and naive, so this feature should not be used with untrusted fonts.

This feature is enabled only if RendererSettings.useAttachedFonts is true. It requires ES6 typed arrays (ArrayBuffer, DataView, Uint8Array, etc).