Fixed FontFace constructor to be consistent between FF and Chrome.
Given a font family with strange characters like ? or [, Chrome is okay with using it as-is `new FontFace("???", "url")` but FF wants it quoted like it would be in a CSS @font-face declaration `new FontFace("\"???\"", "url")`

Originally the code always used the latter form to work on both browsers. However this actually doesn't work on Chrome because Chrome treats that as a separate font family from ???. That is, if a piece of text specifies its CSS font-family as `"???"` FF will match it but Chrome won't. Chrome instead treats `???` and `"???"` as separate font families and would want to match the CSS font family only to the former, not the latter.

The fix is to load both families and expect atleast one to succeed. On FF the unquoted one will fail with a SyntaxError and the quoted one will succeed. On Chrome both will succeed and only the unquoted one will be used by the rendered subs.