Reintroduce `animation-fill-mode: both`
If a subtitle with a fade-out doesn't get removed immediately after its animation end time, it would become completely visible for a brief moment and appear as a flash.

libjass used to use `animation-fill-mode: both` to counter this, but that didn't work in all browsers back then. Fades used to be rendered as two animations (one to fade in and the other to fade out) and some browsers would interpret this to mean that opacity should be 100% after the end time instead of 0%, effectively still causing a flash. So libjass [changed to using only one animation for fades and removed the use of animation-fill-mode.](https://github.com/Arnavion/libjass/commit/72df7fc)

The flash was technically still a problem, but libjass also changed to use a higher-resolution clock based on requestAnimationFrame, so the period between the end of the animation and the sub being removed from the DOM was usually too brief to cause a flash. However low-resolution clocks (250-500ms) like the one used by videojs-ass do cause the flash.

Since fades are rendered as a single animation now, they wouldn't have the problem of being interpreted as 100% opacity past the animation end time, so `animation-fill-mode: both` can be reintroduced to solve the flash problem.

Reported in SunnyLi/videojs-ass#3