Removed fullscreen handling in DefaultRenderer.
It was hacky when it was first written - it required max z-index on the subs wrapper to be visible over the fullscreen video. Now it works on even fewer browsers.

Users who want subs on fullscreen video should implement their own fullscreen functionality to make both the video and the subs wrapper fullscreen.